Serious component movement.
So, when I started this big commit refactor, I started on putting Chat into it's own component first.
That was as mistake -- I should have worked backwards up from the smallest component to the largest.
I may have missed some stuff but for the most part I think I got most styles ported over. Some functionality is still missing.

Here are some key things that happened:

- Putting the "store" into a "<Provider"> as per the redux docs:
http://redux.js.org/docs/basics/UsageWithReact.html#passing-the-store

- Get Chat, Input, Message and Messages into their own components.
- Move old commented components (<Notification> and <Status> into a "TOMAKE" file)
- started removing old code from script.jsx.
