+++ src/components/Chat/Chat_styles.css
@@ -1,0 +1,53 @@
+.Chat-outerWrapper {
+  position: relative;
+  width: 80%;
+  margin: 0 auto;
+}
+
+.Chat-box {
+  position: absolute;
+  right: 0;
+  bottom: 0;
+  background-color: #ffffff;
+}
+
+.Chat-innerWrapper {
+  display: block;
+  width: 280px;
+  height: 360px;
+  box-shadow: #dddddd 1px 1px 8px 0;
+  border-width: 0 1px 1px 0;
+  border-radius: 3px 3px 0 0;
+  border-style: solid;
+  border-color: #cccccc;
+  border-bottom: 0;
+}
+
+.Chat-header {
+  display: block;
+  width: 100%;
+  height: 32px;
+  padding: 8px;
+  box-sizing: border-box;
+  border-radius: 3px 3px 0 0;
+  box-shadow: 0 1px 1px 0 rgba(0, 0, 0, 0.15);
+  font-size: 13px;
+  font-family: \Arial\;
+  sans-serif;
+  color: white;
+  background: #ef7f7f;
+}
+
+.Chat-headerText {}
+
+.Chat-icon {
+  border: 0;
+  float: right;
+  font-size: 24px;
+  margin-right: -8px;
+  margin-top: -7px;
+  font-weight: 800;
+  cursor: pointer;
+  color: rgba(255, 255,255, 0.5);
+  background: rgba(0, 0, 0, 0);
+}
