+++ src/components/Input/Input_styles.css
@@ -1,0 +1,18 @@
+.Input-input {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  width: 100%;
+  height: 48px;
+  box-sizing: border-box;
+  border: 0;
+  border-right: 1px solid #ccc;
+  border-top: 1px solid #ddd;
+  background-color: #fff;
+  padding: 8px;
+  color: #222;
+  font-size: 11px;
+  resize: none;
+  outline: 0;
+}
